requirements.txt
selenium
pytest
pytest-html
pytest-metadata
pytest-xdist
allure-pytest
requests
python-dotenv